# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Chatcart API — WhatsApp-first seller catalog management
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface SendOtpRequest {
  /** Phone number in E.164 format (e.g. +919876543210) */
  phone: string;
}

export interface SendOtpResponse {
  message: string;
  /** Seconds until OTP expires */
  expiresIn: number;
}

export interface VerifyOtpRequest {
  phone: string;
  otp: string;
}

export interface Seller {
  id: number;
  phone: string;
  storeName: string;
  subdomain: string;
  whatsappNumber: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  seller: Seller;
}

export interface UpdateSellerRequest {
  storeName?: string;
  whatsappNumber?: string;
}

export interface Category {
  id: number;
  name: string;
  productCount: number;
}

export interface CreateCategoryRequest {
  name: string;
}

export interface UpdateCategoryRequest {
  name: string;
}

export type ProductStatus = typeof ProductStatus[keyof typeof ProductStatus];


export const ProductStatus = {
  active: 'active',
  out_of_stock: 'out_of_stock',
  hidden: 'hidden',
  deleted: 'deleted',
} as const;

export interface ProductImage {
  id: number;
  url: string;
  displayOrder: number;
}

export interface ProductVariant {
  id: number;
  /** e.g. Size, Color, Stone */
  variantType: string;
  /** Available options for this variant type */
  options: string[];
}

export interface Product {
  id: number;
  name: string;
  description?: string;
  /** Price in INR */
  price: number;
  categoryId?: number;
  categoryName?: string;
  status: ProductStatus;
  stockCount: number;
  showWhenOutOfStock: boolean;
  sortOrder: number;
  images: ProductImage[];
  variants: ProductVariant[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateProductRequest {
  name: string;
  description?: string;
  price: number;
  categoryId?: number;
  stockCount?: number;
}

export interface UpdateProductRequest {
  name?: string;
  description?: string;
  price?: number;
  categoryId?: number;
  stockCount?: number;
  status?: ProductStatus;
  showWhenOutOfStock?: boolean;
}

export type ReorderProductsRequestItemsItem = {
  id: number;
  sortOrder: number;
};

export interface ReorderProductsRequest {
  items: ReorderProductsRequestItemsItem[];
}

export interface AddProductImageRequest {
  /** Object path from the upload endpoint (e.g. /objects/uploads/uuid) */
  url: string;
  displayOrder?: number;
}

export type ReorderImagesRequestItemsItem = {
  id: number;
  displayOrder: number;
};

export interface ReorderImagesRequest {
  items: ReorderImagesRequestItemsItem[];
}

export interface AddProductVariantRequest {
  variantType: string;
  options: string[];
}

export interface UpdateProductVariantRequest {
  variantType?: string;
  options?: string[];
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  fulfilled: 'fulfilled',
} as const;

export interface Order {
  /** e.g. ORD-00001 */
  id: string;
  customerContact?: string;
  status: OrderStatus;
  totalAmount: number;
  itemCount: number;
  createdAt: string;
}

export interface OrderItem {
  id: number;
  productNameSnapshot: string;
  priceSnapshot: number;
  variantSnapshot?: string;
  quantity: number;
}

export type OrderDetail = Order & {
  items: OrderItem[];
};

export interface OrderListResponse {
  orders: Order[];
  total: number;
  page: number;
  limit: number;
}

export interface DashboardStats {
  activeProducts: number;
  outOfStockProducts: number;
  hiddenProducts: number;
  totalOrders: number;
  recentOrders: Order[];
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type ListProductsParams = {
status?: ProductStatus;
categoryId?: number;
search?: string;
};

export type ListOrdersParams = {
page?: number;
limit?: number;
};