# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Chatcart API — WhatsApp-first seller catalog management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/send-otp:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send OTP to phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpRequest"
      responses:
        "200":
          description: OTP sent (or logged to console in dev)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendOtpResponse"
        "400":
          description: Invalid phone number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP and issue session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpRequest"
      responses:
        "200":
          description: Authenticated — returns JWT token and seller profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated seller
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current seller profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Seller"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Invalidate current session
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Seller ────────────────────────────────────────────────────────────────
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories for the authenticated seller
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a new category
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{categoryId}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category name
      security:
        - BearerAuth: []
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryRequest"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      security:
        - BearerAuth: []
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Products ──────────────────────────────────────────────────────────────
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for the authenticated seller
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated order list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order with line items
      security:
        - BearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order detail with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Dashboard ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    # ── Auth ──────────────────────────────────────────────────────────────
    SendOtpRequest:
      type: object
      properties:
        phone:
          type: string
          description: Phone number in E.164 format (e.g. +919876543210)
      required:
        - phone
    SendOtpResponse:
      type: object
      properties:
        message:
          type: string
        expiresIn:
          type: integer
          description: Seconds until OTP expires
      required:
        - message
        - expiresIn
    VerifyOtpRequest:
      type: object
      properties:
        phone:
          type: string
        otp:
          type: string
      required:
        - phone
        - otp
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        seller:
          $ref: "#/components/schemas/Seller"
      required:
        - token
        - seller

    # ── Seller ────────────────────────────────────────────────────────────
    Seller:
      type: object
      properties:
        id:
          type: integer
        phone:
          type: string
        storeName:
          type: string
        subdomain:
          type: string
        whatsappNumber:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - phone
        - storeName
        - subdomain
        - whatsappNumber
        - createdAt
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        productCount:
          type: integer
      required:
        - id
        - name
        - productCount
    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    UpdateCategoryRequest:
      type: object
      properties:
        name:
          type: string
      required:
        - name

    # ── Products ──────────────────────────────────────────────────────────
...
    OrderStatus:
      type: string
      enum:
        - pending
        - confirmed
        - fulfilled
    Order:
      type: object
      properties:
        id:
          type: string
          description: "e.g. ORD-00001"
        customerContact:
          type: string
        status:
          $ref: "#/components/schemas/OrderStatus"
        totalAmount:
          type: number
        itemCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - status
        - totalAmount
        - itemCount
        - createdAt
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productNameSnapshot:
          type: string
        priceSnapshot:
          type: number
        variantSnapshot:
          type: string
        quantity:
          type: integer
      required:
        - id
        - productNameSnapshot
        - priceSnapshot
        - quantity
    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrderItem"
          required:
            - items
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - orders
        - total
        - page
        - limit

    # ── Dashboard ─────────────────────────────────────────────────────────
...